Add ground truth as a block collection
This file contains the duplicate entities.

The first number of a pair refers to the original dataset and
the second number refers to the index (position) in the dataset.

For example:

    [[1, 0], [2, 5]]

Means that the first (0) entity of dataset1 (1) is the same as
the sixth entity (5) of dataset2 (2).

This indexing is also used in the code of this project.